relocate notifications to App.tsx to reduce duplication
