+++ src/App.tsx
@@ -1,11 +1,14 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Redirect } from "react-router-dom";
-import { AppBar, Toolbar, Typography, Button } from "@material-ui/core";
+import { AppBar, Toolbar, Typography, Button, Snackbar, IconButton } from "@material-ui/core";
 import styled from "styled-components";
+import { useDispatch, useSelector } from "react-redux";
+import CloseIcon from "@material-ui/icons/Close";
 
 import { AddArtists, EditArtist } from "./artist";
 import { Dashboard } from "./dashboard";
 import { AddNews } from "./news";
+import { deleteNotificationAction, selectNotification } from "./notifications";
 
 const StyledAppTitle = styled(Typography)`
     flex-grow: 1;
@@ -37,7 +40,14 @@ function App() {
     //         <Button color="inherit">Login</Button>
     //     </>
     // );
-
+    const dispatch = useDispatch();
+    const handleClose = (event: React.SyntheticEvent | React.MouseEvent, reason?: string) => {
+        if (reason === "clickaway") {
+            return;
+        }
+        dispatch(deleteNotificationAction());
+    };
+    const notification = useSelector(selectNotification);
     return (
         <Router>
             <AppBar position="relative">
@@ -57,6 +67,24 @@ function App() {
             <Route exact={true} path="/artists" component={AddArtists} />
             <Route path="/artist/:id" component={EditArtist} />
             <Route exact={true} path="/news" component={AddNews} />
+            <Snackbar
+                anchorOrigin={{
+                    vertical: "bottom",
+                    horizontal: "left",
+                }}
+                open={notification.showNotification}
+                autoHideDuration={10000}
+                onClose={handleClose}
+                ContentProps={{
+                    "aria-describedby": "message-id",
+                }}
+                message={<span id="message-id">{notification.message}</span>}
+                action={[
+                    <IconButton key="close" aria-label="close" color="inherit" onClick={handleClose}>
+                        <CloseIcon />
+                    </IconButton>,
+                ]}
+            />
         </Router>
     );
 }
